+++ crates/weaver-core/src/storage/tables/shared_table.rs
@@ -53,7 +53,7 @@ impl DynamicTable for SharedTable {
         self.0.read(tx, key)
     }
 
-    fn size_estimate(&self, key_index: &KeyIndex) -> Result<usize, WeaverError> {
+    fn size_estimate(&self, key_index: &KeyIndex) -> Result<u64, WeaverError> {
         self.0.size_estimate(key_index)
     }
